pyproject.toml
[tool.poetry]
name = "habit tracker"
version = "0.1.0"
description = ""
authors = ["Mark Pochivalov <pochivalov.mark@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.3.0"
celery = "^5.3.1"
redis = "^5.0.0"
django-celery-beat = "^2.5.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
